Refactor Modal into a simple component with minimal logic and styling, that we can use for any modal that should cover the entire screen. Added a color scheme to Checkbox Field.
