+++ source/components/molecules/index.js
@@ -12,7 +12,6 @@ export { default as Header } from './Header';
 export { default as InputForm } from './InputForm';
 export { default as ListItem } from './ListItem';
 export { default as MaterialTopTabBarWrapper } from './MaterialTopTabBarWrapper';
-export { default as Modal } from './Modal';
 export { default as ScreenWrapper } from './ScreenWrapper';
 export { default as StoryWrapper } from './StoryWrapper';
 export { default as TabBarIcon } from './TabBarIcon';
@@ -30,3 +29,4 @@ export { default as FormField } from './FormField';
 export { default as ButtonField } from './ButtonField';
 export { default as SubstepButton } from './SubstepButton';
 export { default as SubstepModal } from './SubstepModal';
+export { default as Modal } from './Modal';
